Rename test and adjust namespaces
